feat: Separate cell edit and cell display.
