+++ components/ADempiere/FieldDefinition/FieldSearch/GeneralInfoSearch/panel.vue
@@ -78,7 +78,7 @@
       >
         <template slot-scope="scope">
           <!-- formatted displayed value -->
-          <cell-info
+          <cell-display-info
             :parent-uuid="metadata.parentUuid"
             :container-uuid="uuidForm"
             :field-attributes="fieldAttributes"
@@ -137,7 +137,7 @@ import { DISPLAY_COLUMN_PREFIX } from '@/utils/ADempiere/dictionaryUtils'
 
 // components and mixins
 import fieldSearchMixin from '../mixinFieldSearch'
-import CellInfo from '@theme/components/ADempiere/DataTable/Components/CellInfo.vue'
+import CellDisplayInfo from '@theme/components/ADempiere/DataTable/Components/CellDisplayInfo.vue'
 import FieldDefinition from '@theme/components/ADempiere/FieldDefinition/index.vue'
 import CustomPagination from '@theme/components/ADempiere/DataTable/Components/CustomPagination.vue'
 import IndexColumn from '@theme/components/ADempiere/DataTable/Components/IndexColumn.vue'
@@ -151,9 +151,9 @@ export default {
   name: 'PanelGeneralInfoSearch',
 
   components: {
+    CellDisplayInfo,
     CustomPagination,
     FieldDefinition,
-    CellInfo,
     IndexColumn
   },
